add_more_product working when you press + in cart

+++ handlers/public.py
@@ -20,6 +20,18 @@ def about():
     return render_template("public/about.html")
 
 
+def add_more_product(product_id):
+    session_cookie = request.cookies.get("session")
+    product = db.query(Product).get(int(product_id))
+    user = db.query(User).filter_by(session_token=session_cookie).first()
+    cart_item = db.query(Cart).filter_by(user=user, product=product, invoice=None).first()
+
+    cart_item.quantity += 1
+    cart_item.save()
+
+    return redirect(url_for("public.cart", product_id=product_id))
+
+
 def add_to_cart(product_id):
     session_cookie = request.cookies.get("session")
     product = db.query(Product).get(int(product_id))
@@ -100,7 +112,10 @@ def delete_cart_product(product_id):
         product = db.query(Product).get(int(product_id))
         user = db.query(User).filter_by(session_token=session_cookie).first()
         cart_item = db.query(Cart).filter_by(user=user, product=product, invoice=None).first()
-
-        cart_item.delete()
+        # urediti da ce je produkt zbrisan s strani admina da pokaze template
+        if product:
+            cart_item.delete()
+        else:
+            return "This product does not exist"
 
         return redirect(url_for("public.cart"))
